server/requirements.txt
aiohttp
PyJWT
bcrypt
